requirements.txt
pymysql==1.1.0
python-dotenv==1.0.0
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.31.0
gunicorn==21.2.0
logging==0.4.9.6 